pyproject.toml
[project]
name = "python-starter"
version = "0.1.0"
description = "Python Playground Kit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce==2.5.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "justsdk>=2.2.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--verbose",
    "--no-header",
    "-s"
]
